fix long database not int as id

+++ mad-schuelerturnier-web/src/main/java/com/googlecode/madschuelerturnier/web/controllers/DBFileUploadController.java
@@ -38,7 +38,7 @@ public class DBFileUploadController implements Serializable {
 
     public void handleFileUpload(FileUploadEvent event) {
 
-        File file = repo.findByTypAndPearID(typeToSet, Integer.parseInt(idToMatchWith));
+        File file = repo.findByTypAndPearID(typeToSet, Long.parseLong(idToMatchWith));
 
         if (file == null) {
             file = new File();
